cmake: Switch suitable PUBLIC target_link_libraries to PRIVATE.
Declaring a library as PUBLIC in target_link_libraries tells downstream
linkers of the target to also need to link to the PUBLIC libraries. This
is necessary when the headers of those libraries are required by the
public headers of the target library. There were some instances where
linked libraries were declared PUBLIC but were not required to use the
public headers, so they have been switched to PRIVATE declarations. This
will allow OOT modules to reduce their linking and dependencies in some
cases.

Signed-off-by: Ryan Volz <ryan.volz@gmail.com>